pyproject.toml
[project]
name = "bluediamond-api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yousuf", email = "yrehan67@gmail.com" }
]
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bluediamond_api"]
